3.12 Class Node changed to nested class of IntLinkedList
