+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -40,4 +40,18 @@ public class IntLinkedList {
         return ret;
     }
 
+    private class Node {
+
+        private final int value;
+        public Node prev, next;
+
+        public Node(final int i) {
+            value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+    }
 }
